requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
langgraph
chromadb
pypdf
pymupdf
tiktoken
python-dotenv
requests
beautifulsoup4 